Move community stakes and transactions history to profile
