+++ packages/commonwealth/client/scripts/views/components/Profile/ProfileActivity/TransactionsTab/NoTransactionHistory/NoTransactionHistory.scss
@@ -1,4 +1,4 @@
-@import '../../../../styles/shared';
+@import '../../../../../../styles/shared.scss';
 
 .NoTransactionHistory {
   width: 100%;
