+++ packages/commonwealth/client/scripts/views/components/Profile/ProfileActivity/TransactionsTab/Stakes/Stakes.scss
@@ -1,4 +1,4 @@
-@import '../../../../styles/shared';
+@import '../../../../../../styles/shared.scss';
 
 .Stakes {
   .Table {
