+++ packages/commonwealth/client/scripts/views/components/Profile/ProfileActivity/TransactionsTab/Stakes/index.ts
