+++ packages/commonwealth/client/scripts/views/components/Profile/ProfileActivity/TransactionsTab/TransactionHistory/TransactionHistory.scss
@@ -1,4 +1,4 @@
-@import '../../../../styles/shared';
+@import '../../../../../../styles/shared.scss';
 
 .Transactions {
   .Table {
