+++ packages/commonwealth/client/scripts/views/components/Profile/ProfileActivity/TransactionsTab/TransactionHistory/index.ts
@@ -1,0 +1,1 @@
+export { TransactionHistory as default } from './TransactionHistory';
