+++ packages/commonwealth/client/scripts/views/components/Profile/ProfileActivity/TransactionsTab/TransactionsTab.scss
@@ -1,6 +1,6 @@
-@import '../../../styles/shared';
+@import '../../../../../styles/shared';
 
-.MyCommunityStake {
+.TransactionsTab {
   display: flex;
   flex-direction: column;
   width: 100%;
