+++ packages/commonwealth/client/scripts/views/components/Profile/ProfileActivity/TransactionsTab/index.ts
@@ -1,0 +1,1 @@
+export { TransactionsTab } from './TransactionsTab';
