--- packages/commonwealth/client/scripts/views/pages/MyCommunityStake/Transactions/index.ts
@@ -1,1 +1,0 @@
-export { Transactions as default } from './Transactions';
