--- packages/commonwealth/client/scripts/views/pages/MyCommunityStake/index.ts
@@ -1,1 +1,0 @@
-export { MyCommunityStake as default } from './MyCommunityStake';
